+++ src/api/admin/mpn/automations/available-actions/route.ts
@@ -3,7 +3,7 @@ import {
   MedusaResponse,
 } from "@medusajs/framework/http"
 import { MPN_AUTOMATION_MODULE } from "../../../../../modules/mpn-automation"
-import { MpnAutomationService } from "../../../../../modules/mpn-automation/services"
+import MpnAutomationService from "../../../../../modules/mpn-automation/service"
 
 export async function GET(
   req: MedusaStoreRequest,
@@ -18,10 +18,34 @@ export async function GET(
     | string
     | undefined
 
-  const actions =
-    automationService.getAvailableActions(eventName)
+  const getActionHandlers =
+    automationService.getActionHandlers()
+
+  const getActionHandlersMapped = await Promise.all(
+    Array.from(getActionHandlers.entries()).map(
+      async ([key, value]) => {
+        {
+          const action = value.handler
+          const enabled = value.enabled
+          let data = {}
+  
+          if ((action as any)?.fetchData) {
+            data = await (action as any)?.fetchData({
+              container: req.scope,
+              eventName: eventName,
+            })
+          }
+  
+          return {
+            ...data,
+            enabled: enabled,
+          }
+        }
+      }
+    )
+  )
 
   res.json({
-    actions: actions,
+    actions: getActionHandlersMapped,
   })
 }