+++ src/subscribers/mpn.automation.action.slack.executed.ts
@@ -23,14 +23,19 @@ export default async function mpnAutomationActionSlackExecutedHandler({
   container,
 }: SubscriberArgs<any>) {
   const {
+    eventName: triggerEventName,
     action,
     context,
-    eventName: triggerEventName,
     contextType,
   } = data
   const config = container.resolve("configModule") as any
   const moduleConfig = config?.modules.mpnAutomation
   const backendUrl = moduleConfig?.options.backend_url
+
+  const contextData = {
+    ...context,
+    backendUrl: backendUrl,
+  }
 
   // Execute slack action workflow
   const { result } = await runSlackActionWorkflow(
@@ -40,12 +45,10 @@ export default async function mpnAutomationActionSlackExecutedHandler({
       action: {
         ...action,
         config: {
-          ...action.config,
-          template: action.config.templateName,
-          backendUrl: backendUrl,
+          ...action.config
         },
       },
-      context: context,
+      context: contextData,
       contextType: contextType,
       eventName: triggerEventName,
     },